[KJots] Port bookmarks code, unify link handling code
Summary:
Link creation was present in three places:

 1) Inside template `pagetemplate.html`
 2) Inside `KJotsLinkDialog`, the dialog to add links
 3) Inside `KJotsBookmarks`

Unify it inside `KJotsModel`: provide a `UrlRole` for `QModelIndex::data`.
Provide an `url` property for `KJotsEntity`, so it could also be used in templates.

Link handling was also duplicated in two places:

 1) Inside `KJotsBookmarks`
 2) Inside `KJotsBrowser`

Unify it too, so both classes emit a signal with `QUrl` property, which is then caught by `KJotsWidget`
(which actually gets the job done)

Test Plan:
 1) Navigation in the "Browser mode" works, clicking at "books", "pages" and "page headers" (which should switch to edit mode)
 2) Bookmarks also work, they are created with fancy icons, and navigation works (both with "pages" and "books")
 3) Link dialog creates proper links pointing where they should
 4) External links are being opened by an external web browser

Reviewers: dvratil

Reviewed By: dvratil

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D29073