[KJotsEdit] Reimplement ancient (yet still relevant) workaround for QTextEdit
Summary:
If the very first block of the document has some style, the cursor is not aware of it.
However, explicitly telling to move the cursor to the start of the document makes the cursor catch up the style
(it's probably slighly better than moving cursor to the left / to the right).

The best I could find was this stackoveflow, unfortunately:
https://stackoverflow.com/questions/37160039/set-default-character-format-in-qtextdocument

Test Plan:
1) Make a document starting with a, say, bold word
2) Close && Reopen KJots (so that it doesn't reload already used cursor)
3) Reopen this document -> the cursor is shown as bold from the very beginning

Reviewers: dvratil, dfaure

Reviewed By: dfaure

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D29107