[KJots] Add support for headings
Summary:
This patch adds support for headings, introduced for underlying `KRichTextEdit` in {D28854}.
It adds relevant actions to toolbars & menu.
This also increases the required KF5 version to (not yet released) 5.70.

Slighly overlaps with {D29291} (which increases the version in .rc-file)

CCBUG: 230317

Test Plan: it compiles, actions are added (and functional)

Reviewers: dvratil, mlaurent

Subscribers: mlaurent, kde-pim

Differential Revision: https://phabricator.kde.org/D29300