Port KJots to KPimTextEdit
This includes using RichTextComposer instead of KRichTextWidget for KJotsEdit,
and writing an own wrapper around KJotsBrowser with Text-To-Speech widget and Find Bar.

It has following benefits for free:
 - Inline find / replace bar (instead of having dialog & a lot of custom code for that)
 - Text-To-Speech integration
 - Add Image dialog
 - Action to paste without formatting

The only feature that is lost is search-in-all-documents.
It's something that should be done properply (e.g. indexing, Xapian, etc).
As a workaround, one can open preview for whole bookshelf and use Find there...